res worked around a possible VC7 64bit warning in   csIntegralHashKeyHandler<>.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18347 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/hash.cpp
@@ -18,7 +18,7 @@
 #include "cssysdef.h"
 #include "csutil/hash.h"
 
-unsigned int csHashCompute (char const* s, int n)
+unsigned int csHashCompute (char const* s, size_t n)
 {
   unsigned int h = 0;
   const char* end = s + n;
